Refactor data getting functionality: remove files 'config.js' and 'page.js'
