--- js/config.js
@@ -1,9 +1,0 @@
-const SIMILAR_ANNOUNCEMENTS_COUNT = 10;
-const GET_ANNOUNCEMENTS_URL = 'https://25.javascript.pages.academy/keksobooking/data';
-const SEND_DATA_URL = 'https://25.javascript.pages.academy/keksobooking';
-
-export {
-  SIMILAR_ANNOUNCEMENTS_COUNT,
-  GET_ANNOUNCEMENTS_URL,
-  SEND_DATA_URL,
-};
