--- js/page.js
@@ -1,9 +1,0 @@
-const showDataNotLoadedError = (errorMessage) => {
-  const mapElement = document.querySelector('.map');
-  const errorElement = document.createElement('div');
-  errorElement.classList.add('map__error');
-  errorElement.textContent = errorMessage;
-  mapElement.appendChild(errorElement);
-};
-
-export {showDataNotLoadedError};
